Added a new disassembler plugin, DisassemblerLLVMC, which uses the Disassembler.h interface to the LLVM disassemblers rather than the EnhancedDisassembly.h interface.  Disassembler.h is a better-maintained API and will be stabler in the long term.
Currently the output from Disassembler.h does not
provide for symbolic disassembly in all the places
that the old disassembler did, so I have gated (and
disabled) the disassembler.  It'll be easy to flip
the switch later.

In the meantime, to enable the new disassembler,
uncomment "#define USE_NEW_DISASSEMBLER" in
lldb.cpp.

llvm-svn=150772